refactor: combine filters into 1 component
