+++ packages/shared/components/atoms/iconButtons/index.js
@@ -1,2 +1,1 @@
 export { default as SyncSelectedAccountIconButton } from './SyncSelectedAccountIconButton.svelte'
-export { default as FilterButton } from '../../filters/ActivityFilter.svelte'
