+++ packages/shared/components/filters/index.js
@@ -1,3 +1,2 @@
-export { default as ActivityFilter } from './ActivityFilter.svelte'
-export { default as AssetFilter } from './AssetFilter.svelte'
+export { default as Filter } from './Filter.svelte'
 export { default as FilterItem } from './FilterItem.svelte'
