+++ packages/shared/lib/core/wallet/interfaces/filter.interface.ts
@@ -1,12 +1,14 @@
+export type Filter = ActivityFilter | AssetFilter
+
 export interface ActivityFilter {
     showHidden: BooleanFilter
 }
 
-export type Filter = BooleanFilter
-
 export interface AssetFilter {
     showHidden: BooleanFilter
 }
+
+export type ValueFilter = BooleanFilter
 
 export type BooleanFilter = {
     type: 'boolean'
